Update peripheral.js
If the device does not respond with at least the # of requested services, bail early. This keeps the user from relying on a timeout.

The way this is currently written, if the amount of services returned does not match the request, it never calls the callback. It calls the callback when it has processed the same # of services as the request, with no guarantee they are the correct ones requested, and ignoring any extra ones. 

This should also probably have an error path after the for..in loop in case numDiscovered was insufficient, but baby steps.